Add user detail component
