+++ src/app/profile/page.tsx
@@ -1,36 +1,18 @@
 'use client';
 
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { useSelector } from 'react-redux';
-import { useSearchParams } from 'next/navigation';
 
 import { RootState, useAppDispatch } from 'store/store';
 import { fetchAuthenticatedUser } from 'store/slices/singleUserSlice';
 
-import { Repositories } from 'app/users/Repositories';
-
-import { Followers } from 'components/Followers';
-import { Following } from 'components/Following';
-import { Starred } from 'components/Starred';
 import { Spinner } from 'components';
-import { UserDetailTabs } from 'components/UserDetailTabs';
+
 import { UserDetailInformation } from 'components/UserDetailInformation';
 import { UserDetailHeader } from 'components/UserDetailHeader';
+import { UserDetail } from 'app/profile/UserDetail';
 
 const Details = () => {
-  const searchParams = useSearchParams();
-
-  const [selectedTab, setSelectedTab] = useState(
-    searchParams.get('tab') || 'followers'
-  );
-
-  useEffect(() => {
-    const tab = searchParams.get('tab');
-    if (tab && tab !== selectedTab) {
-      setSelectedTab(tab);
-    }
-  }, [searchParams, selectedTab]);
-
   const dispatch = useAppDispatch();
 
   const { status } = useSelector((state: RootState) => state.singleUser);
@@ -51,15 +33,7 @@ const Details = () => {
     <>
       <UserDetailHeader />
       <UserDetailInformation />
-      <UserDetailTabs
-        selectedTab={selectedTab}
-        setSelectedTab={(tab) => setSelectedTab(tab)}
-      />
-
-      {selectedTab === 'followers' && <Followers status={status} />}
-      {selectedTab === 'following' && <Following status={status} />}
-      {selectedTab === 'starred' && <Starred status={status} />}
-      {selectedTab === 'repos' && <Repositories status={status} />}
+      <UserDetail />
     </>
   );
 };
